server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
openai==1.58.1
httpx==0.27.2
python-dotenv==1.0.1